fixed post_func.js and post.js naming
