+++ frontend_soundgood/src/App.js
@@ -8,7 +8,7 @@ import DiscoveredSongs from './Pages/DiscoveredSongs'
 import Landing from './Pages/Landing'
 import Register from './Pages/Register.js'
 import Login from './Pages/Login.js'
-import Post_func from './Pages/Post_func'
+import Post from './Pages/Post'
 
 function App() {
   const isLoggedIn = () => {
@@ -25,7 +25,7 @@ function App() {
             <Route path='/savedSongs' element={<SavedSongs />} />
             <Route path='/forum' element={<Forum />} />
             <Route path='/discoveredSongs' element={<DiscoveredSongs />} />
-            <Route path='/post' element={<Post_func />} />
+            <Route path='/post' element={<Post />} />
           </Routes>
         </div>
       ) : (
